video-streaming-app/requirements.txt
google-api-python-client
google-cloud-pubsub
google-cloud-storage==1.6.0
numpy
pandas
gunicorn==19.7.1
Flask==0.12.2
apache-beam[gcp]